Rewrite pkg `scanner` to `fsprov` and use provider interface
