+++ internal/dataprovider/fsprov/parse.go
@@ -1,4 +1,4 @@
-package scanner
+package fsprov
 
 import (
 	"regexp"
